Tic-Tac-Toe
This is a classic Tic Tac Toe (X/O) game built using React.js. It is a two-player game played on a 3x3 grid, where players take turns marking the spaces with X or O. The first player to align three of their marks in a row, column, or diagonal wins.